chore: move viewport-manager, reshuffle preferences
