+++ packages/autopilot/src/app/controllers/layout.ts
@@ -15,7 +15,7 @@
 import { util } from '@automationcloud/engine';
 import { inject, injectable } from 'inversify';
 import { controller } from '../controller';
-import { ViewportManager } from '../managers/viewport-manager';
+import { ViewportManager } from '../viewports/viewport-manager';
 import { UserData } from '../userdata';
 import { LayoutDirection, LayoutItem } from '../util/layout-item';
 import { Viewport } from '../viewport';
