+++ packages/autopilot/src/app/viewports/viewport-manager.ts
